Using a HttpURLConnectionClient to create the appropriate client.
