+++ library/src/main/java/com/parse/internal/signpost/basic/NativeHttpURLConnectionClient.java
@@ -1,0 +1,14 @@
+package com.parse.internal.signpost.basic;
+
+import java.net.HttpURLConnection;
+import java.net.URL;
+
+/**
+ * Adapter class to provide HttpURLConnection, using android HttpURLConnection or OkHttp
+ */
+public final class NativeHttpURLConnectionClient implements HttpURLConnectionClient {
+
+    public HttpURLConnection open(URL url) throws Exception {
+        return (HttpURLConnection) url.openConnection();
+    }
+}
